// Enterprise-Multi-Layer-Scheduling-System/scheduler-frontend/src/types/org.ts
export interface SubTeam {
  id: string;
  name: string;
  description?: string;
  timezone?: string;
  parent_team_id: string;

  lead?: {
    id: string;
    first_name: string;
    last_name: string;
  };

  _count?: {
    sub_team_members: number;
  };
}